[NSE-772] Code refactor for ColumnarBatchScan (#805)
* Override doCanonicalize in ColumnarBatchScanExec

* Refactor ColumnarBatchScan
