+++ native-sql-engine/core/src/main/scala/com/intel/oap/extension/ColumnarOverrides.scala
@@ -27,7 +27,6 @@ import com.intel.oap.sql.shims.SparkShimLoader
 
 import org.apache.spark.{MapOutputStatistics, SparkContext}
 import org.apache.spark.internal.Logging
-import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.{SparkSession, SparkSessionExtensions}
 import org.apache.spark.sql.catalyst.expressions._
 import org.apache.spark.sql.catalyst.optimizer.BuildLeft
@@ -51,7 +50,6 @@ import org.apache.spark.sql.execution.joins._
 import org.apache.spark.sql.execution.python.{ArrowEvalPythonExec, ColumnarArrowEvalPythonExec}
 import org.apache.spark.sql.execution.window.WindowExec
 import org.apache.spark.sql.internal.SQLConf
-import org.apache.spark.sql.vectorized.ColumnarBatch
 
 import org.apache.spark.util.ShufflePartitionUtils
 
@@ -90,26 +88,7 @@ case class ColumnarPreOverrides() extends Rule[SparkPlan] {
     case plan: BatchScanExec =>
       logDebug(s"Columnar Processing for ${plan.getClass} is currently supported.")
       val runtimeFilters = SparkShimLoader.getSparkShims.getRuntimeFilters(plan)
-      new ColumnarBatchScanExec(plan.output, plan.scan, runtimeFilters) {
-        // This method is a commonly shared implementation for ColumnarBatchScanExec.
-        // We move it outside of shim layer to break the cyclic dependency caused by
-        // ColumnarDataSourceRDD.
-        override def doExecuteColumnar(): RDD[ColumnarBatch] = {
-          val numOutputRows = longMetric("numOutputRows")
-          val numInputBatches = longMetric("numInputBatches")
-          val numOutputBatches = longMetric("numOutputBatches")
-          val scanTime = longMetric("scanTime")
-          val inputSize = longMetric("inputSize")
-          val inputColumnarRDD =
-            new ColumnarDataSourceRDD(sparkContext, partitions, readerFactory,
-              true, scanTime, numInputBatches, inputSize, tmpDir)
-          inputColumnarRDD.map { r =>
-            numOutputRows += r.numRows()
-            numOutputBatches += 1
-            r
-          }
-        }
-      }
+      new ColumnarBatchScanExec(plan.output, plan.scan, runtimeFilters)
     case plan: CoalesceExec =>
       ColumnarCoalesceExec(plan.numPartitions, replaceWithColumnarPlan(plan.child))
     case plan: InMemoryTableScanExec =>
