+++ native-sql-engine/core/src/main/scala/com/intel/oap/extension/columnar/ColumnarGuardRule.scala
@@ -37,7 +37,6 @@ import org.apache.spark.sql.execution.joins._
 import org.apache.spark.sql.execution.python.ArrowEvalPythonExec
 import org.apache.spark.sql.execution.python.ColumnarArrowEvalPythonExec
 import org.apache.spark.sql.execution.window.WindowExec
-import org.apache.spark.sql.vectorized.ColumnarBatch
 
 case class RowGuard(child: SparkPlan) extends SparkPlan {
   def output: Seq[Attribute] = child.output
@@ -81,26 +80,7 @@ case class ColumnarGuardRule() extends Rule[SparkPlan] {
         case plan: BatchScanExec =>
           if (!enableColumnarBatchScan) return false
           val runtimeFilters = SparkShimLoader.getSparkShims.getRuntimeFilters(plan)
-          new ColumnarBatchScanExec(plan.output, plan.scan, runtimeFilters) {
-            // This method is a commonly shared implementation for ColumnarBatchScanExec.
-            // We move it outside of shim layer to break the cyclic dependency caused by
-            // ColumnarDataSourceRDD.
-            override def doExecuteColumnar(): RDD[ColumnarBatch] = {
-              val numOutputRows = longMetric("numOutputRows")
-              val numInputBatches = longMetric("numInputBatches")
-              val numOutputBatches = longMetric("numOutputBatches")
-              val scanTime = longMetric("scanTime")
-              val inputSize = longMetric("inputSize")
-              val inputColumnarRDD =
-                new ColumnarDataSourceRDD(sparkContext, partitions, readerFactory,
-                  true, scanTime, numInputBatches, inputSize, tmpDir)
-              inputColumnarRDD.map { r =>
-                numOutputRows += r.numRows()
-                numOutputBatches += 1
-                r
-              }
-            }
-          }
+          new ColumnarBatchScanExec(plan.output, plan.scan, runtimeFilters)
         case plan: FileSourceScanExec =>
           if (plan.supportsColumnar) {
             return false
